--- src/components/screen/CreateTransactionStack/LeftButton.tsx
@@ -1,15 +1,0 @@
-import Button from 'components/@common/Buttons/Button';
-import React from 'react';
-
-type LeftButtonProps = {
-  label?: '이전';
-  onPress: () => void;
-};
-
-export default function LeftButton({label = '이전', onPress}: LeftButtonProps) {
-  return (
-    <Button variant="secondary" onPress={onPress}>
-      {label}
-    </Button>
-  );
-}
